src/main/webapp/WEB-INF/themes/from-plos/test/requirements.txt
appdirs==1.4.3
Appium-Python-Client==0.31
beautifulsoup4==4.8.1
browsermob-proxy==0.8.0
coloredlogs==10.0
config==0.4.2
enum34==1.1.6
flake8==3.7.8
jsonpath==0.82
MarkupSafe==1.0
mysql-connector==2.2.9
packaging==19.2
pyparsing==2.4.2
pytest==5.2.1
pytest-flake8==1.0.4
pytest-html==2.0.0
pytest-repeat==0.8.0
pytest-rerunfailures==7.0
pytest-sugar==0.9.2
pytest-timeout==1.3.3
pytest-xdist==1.30.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
selenium==3.14.1
setuptools==41.4.0
smart_getenv==1.1.0
teamcity-messages==1.25
urllib3==1.25.6
Wand==0.4.4
wget==3.2
wheel==0.33.6
-e git://github.com/testlabauto/loremipsum.git#egg=loremipsum
